human_image_validator/requirements.txt
# requirements.txt
django==5.0.1
djangorestframework==3.14.0
opencv-python==4.8.1
mediapipe==0.10.9
Pillow==10.2.0
numpy==1.26.3
face-recognition==1.3.0
